refactor: harden route load balancer and quiet bad-addr logging (#38)
Findings from a review of the route module:

- RRLB.Get: take the round-robin modulo in uint32 space. int(uint32) can go
  negative on 32-bit platforms once the counter passes MaxInt32, which would
  produce a negative slice index and panic. No effect on amd64 (the build
  target), but correct everywhere now.

- badAddrTable.MarkBad: log the "mark bad" warning only on the transition into
  a bad state instead of on every failed dial. A sustained backend outage at
  high RPS re-marks the same host constantly; logging each one floods the log
  exactly when it's busiest. A recovered-then-failed host (entry expired) logs
  again, so distinct outage episodes stay visible. The store/expiry logic is
  unchanged. Extracted the transition decision into a testable mark() helper.

- Drop the leftover BenchmarkSprintf/StringConcat/StringsJoin micro-benchmarks
  in table_test.go; they exercised no route code (they justified Lookup's
  string concat once and have served their purpose).

- Add coverage for the previously-untested bad-address expiry path: IsBad
  returning false after badDuration, Clear removing only expired entries, and
  the new mark() transition logic (using backdated timestamps, no waiting).

+++ route/rrlb.go
@@ -19,7 +19,9 @@ func (lb *RRLB) Get(badAddr *badAddrTable) (ip string) {
 		return lb.IPs[0]
 	}
 
-	p := int(atomic.AddUint32(&lb.current, 1)) % l
+	// take the modulo in uint32 space: int(uint32) can be negative on 32-bit
+	// platforms once current exceeds MaxInt32, which would yield a negative index.
+	p := int(atomic.AddUint32(&lb.current, 1) % uint32(l))
 	for k := 0; k < l; k++ { // try gets not bad address
 		i := (p + k) % l
 		ip = lb.IPs[i]
